+++ lib/flutter_resizable_container.dart
@@ -1,6 +1,7 @@
 library flutter_resizable_container;
 
 export 'src/resizable_container.dart';
-export 'src/resizable_controller.dart';
+export 'src/resizable_controller.dart' show ResizableController;
 export 'src/resizable_child.dart';
 export 'src/resizable_divider.dart';
+export 'src/resizable_size.dart' show ResizableSize;